+++ fabric/src/main/java/io/github/sakurawald/fuji/core/command/argument/wrapper/impl/Duration.java
@@ -1,0 +1,9 @@
+package io.github.sakurawald.fuji.core.command.argument.wrapper.impl;
+
+import io.github.sakurawald.fuji.core.command.argument.wrapper.abst.StringValue;
+
+public class Duration extends StringValue {
+    public Duration(String value) {
+        super(value);
+    }
+}