+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [command_cooldown] feature: now requires the `confirmation` when using `/command-cooldown delete` command.
+- [core] improve the `command suggestion` and `command validation` for `duration argument type`.